backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
sse-starlette>=1.8.2
python-dotenv>=1.0.0
